backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
psycopg2-binary==2.9.5
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==1.10.7
python-dotenv==1.0.1
alembic==1.13.1
email-validator==2.1.1
bcrypt==4.1.2
pytest==8.2.1
httpx==0.27.0